Add files via upload
This firmware version is based on WB2CBA's ADX_CAT_1.6 firmware but has been augmented to support 7 bands. If you wish to change the band allocation you will need to change the band definition in lines 178-184 of the firmware before uploading to your ADX.  The standard band configuration in this firmware from band 1 to 7 is: 40, 20, 15, 10, 80, 30, 17. This reflects the sequence shown in the video referenced in the read.me file.