Add files via upload
This file contains all the worksheets necessary to record orders, calculate the number of LPF and mainboard components required to build the number of kits required. The sheet contains four tabs: 
- ADX Orders:  Here you enter the details for each kit ordered
- ADX Quad BOM: This sheet calculates the number of individual componts (by value) required to build the number of required kits as well as providing a worksheet to manage and record the orders placed, expected delivery date and number received. The sheet also provides the list of vendors where the Cowtown project purchased the components for their build.
- LPF Details: This tab calculates the number and value of capacitors and toroids required to complete the number of LPF's ordered on the ADX order tab. The sheet also lists the mouser part numbers for SMD capacitors for the LPF's.
LPF Build and Kitting LIst:  This tab can be used as a kitting list check and calculates the length of magnet wire required to build the LPF's ordered with each kit.
LIT BOM & Checklist: This tab can be used as a checklist during the kitting process.